docker/requirements_bloat.txt
actionlib==1.12.1
addict==2.4.0
angles==1.9.12
anyio==3.6.2
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asn1crypto==0.24.0
async-generator==1.10
attrs==22.2.0
Babel==2.11.0
backcall==0.2.0
bleach==4.1.0
bondpy==1.8.5
Brlapi==0.6.6
cached-property==1.5.2
camera_calibration==1.15.2
camera_calibration_parsers==1.11.13
catkin==0.7.29
certifi==2018.1.18
cffi==1.15.1
chardet==3.0.4
click==6.7
colorama==0.3.7
comm==0.1.4
command-not-found==0.3
contextvars==2.4
controller_manager==0.18.4
controller_manager_msgs==0.18.4
cryptography==2.1.4
cupshelpers==1.0
cv_bridge==1.13.1
cycler==0.11.0
Cython==0.29.15
dataclasses==0.8
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
deprecation==2.1.0
diagnostic_analysis==1.9.7
diagnostic_common_diagnostics==1.9.7
diagnostic_updater==1.9.7
distro-info===0.18ubuntu0.18.04.1
dynamic_reconfigure==1.6.5
entrypoints==0.4
gazebo_plugins==2.8.7
gazebo_ros==2.8.7
gencpp==0.6.6
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.17
gennodejs==2.0.1
genpy==0.6.16
h5py==3.1.0
httplib2==0.9.2
idna==3.4
image_geometry==1.13.1
immutables==0.19
importlib-metadata==4.8.3
importlib-resources==5.4.0
interactive-markers==1.11.5
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
ipywidgets==7.8.0
jedi==0.17.2
Jinja2==3.0.3
joblib==1.1.1
joint_state_publisher==1.12.15
json5==0.9.14
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyter-packaging==0.10.6
jupyter-server==1.13.1
jupyterlab==3.2.9
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
jupyterlab-widgets==1.1.5
kdl-parser-py==1.13.3
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.3.1
language-selector==0.1
laser_geometry==1.6.7
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
linecache2==1.0.0
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==2.0.1
matplotlib==3.3.4
message_filters==1.14.13
mistune==0.8.4
nbclassic==0.3.5
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.7
netifaces==0.10.4
notebook==6.4.10
numpy==1.19.5
oauth==1.0.1
olefile==0.45.1
open3d==0.15.2
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
parso==0.7.1
pbr==5.4.5
pexpect==4.2.1
pickleshare==0.7.5
Pillow==8.4.0
prometheus-client==0.17.1
prompt-toolkit==3.0.36
protobuf==3.0.0
ptyprocess==0.7.0
pycairo==1.16.2
pycparser==2.21
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.14.0
PyGObject==3.26.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==3.1.1
pyquaternion==0.9.9
pyRFC3339==1.0
pyrsistent==0.18.0
python-apt==1.6.5+ubuntu0.7
python-dateutil==2.8.2
python-debian==0.1.32
python_qt_binding==0.4.4
pytz==2018.3
pyxdg==0.25
PyYAML==6.0.1
pyzmq==25.1.1
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
resource_retriever==1.12.7
rosbag==1.14.13
rosboost-cfg==1.14.9
rosclean==1.14.9
roscreate==1.14.9
rosgraph==1.14.13
roslaunch==1.14.13
roslib==1.14.9
roslint==0.11.2
roslz4==1.14.13
rosmake==1.14.9
rosmaster==1.14.13
rosmsg==1.14.13
rosnode==1.14.13
rosparam==1.14.13
rospy==1.14.13
rosservice==1.14.13
rostest==1.14.13
rostopic==1.14.13
rosunit==1.14.9
roswtf==1.14.13
rqt-image-view==0.4.17
rqt-moveit==0.5.10
rqt-reconfigure==0.5.4
rqt-robot-monitor==0.5.14
rqt-rviz==0.7.0
rqt_action==0.4.9
rqt_bag==0.5.1
rqt_bag_plugins==0.5.1
rqt_console==0.4.9
rqt_dep==0.4.9
rqt_graph==0.4.11
rqt_gui==0.5.3
rqt_gui_py==0.5.3
rqt_launch==0.4.8
rqt_logger_level==0.4.8
rqt_msg==0.4.8
rqt_nav_view==0.5.7
rqt_plot==0.4.13
rqt_pose_view==0.5.8
rqt_publisher==0.4.8
rqt_py_common==0.5.3
rqt_py_console==0.4.8
rqt_robot_dashboard==0.5.7
rqt_robot_steering==0.5.10
rqt_runtime_monitor==0.5.7
rqt_service_caller==0.4.8
rqt_shell==0.4.9
rqt_srv==0.4.8
rqt_tf_tree==0.6.0
rqt_top==0.4.8
rqt_topic==0.4.11
rqt_web==0.4.8
rviz==1.13.30
scikit-learn==0.24.2
scipy==1.5.4
SecretStorage==2.3.1
Send2Trash==1.8.2
sensor-msgs==1.12.8
simplejson==3.13.2
six==1.11.0
smach==2.0.3
smach_ros==2.0.3
smclib==1.8.5
sniffio==1.2.0
system-service==0.3
systemd-python==234
terminado==0.12.1
testpath==0.6.0
testresources==2.0.0
tf==1.12.1
tf2_geometry_msgs==0.6.5
tf2_kdl==0.6.5
tf2_py==0.6.5
tf2_ros==0.6.5
tf_conversions==1.12.1
threadpoolctl==3.1.0
tomlkit==0.11.6
topic_tools==1.14.13
torch==1.10.2+cpu
torchaudio==0.10.2+cpu
torchvision==0.11.3+cpu
tornado==6.1
tqdm==4.64.1
traceback2==1.4.0
traitlets==4.3.3
typing_extensions==4.1.1
ubuntu-advantage-tools==27.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
unittest2==1.1.0
urdfdom-py==0.4.6
urllib3==1.22
usb-creator==0.3.3
wadllib==1.3.2
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.3.1
widgetsnbextension==3.6.5
xacro==1.13.20
xkit==0.0.0
zipp==3.6.0
zope.interface==4.3.2
